Specify base 10 to the strtol() call converting the device number
The USB device number given as argument to the '-d' option is parsed
with a strtol() call with 0 base.  This is an issue with USB device
numbers that start with a 0 (and they do tend to start with a 0, e.g.
from lsusb output: "Bus 001 Device 014"), because strtol() with 0 base
interprets them as an octal number.  Consequently, the program might not
find the device, or, worse, might find and operate on the wrong device.

Fix this by explicitly specifying base 10 to the strtol() call
converting the device number.

All other strtol() calls already explicitly specify a 10 or 16 base.